+++ changelog.md
@@ -1,3 +1,6 @@
+## Lockout v0.11.2
+- Fixed Entity Counter bug that occurs in 1.21.11
+- Debug Menu compatibility with Sodium
 ## Lockout v0.11.1
 - Fixed GUI inconsistencies
 - Updated Debug Menu to be more consistent across player changes